requirements.txt
absl-py==2.1.0
array_record==0.5.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dicttoxml==1.7.16
dm-tree==0.1.8
docstring_parser==0.16
etils==1.9.4
flatbuffers==24.3.25
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
idna==3.10
immutabledict==4.2.0
importlib_resources==6.4.5
kagglehub==0.3.0
keras==3.5.0
keras-core==0.1.7
keras-cv==0.9.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
promise==2.3
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
Pygments==2.18.0
PySocks==1.7.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
selenium==4.25.0
setuptools==75.1.0
simple-parsing==0.1.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-metadata==1.15.0
termcolor==2.4.0
toml==0.10.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.2
